+++ EnvVars.hpp
@@ -28,7 +28,7 @@ THE SOFTWARE.
 #include <time.h>
 #include "Kernels.hpp"
 
-#define TB_VERSION "1.11"
+#define TB_VERSION "1.12"
 
 extern char const MemTypeStr[];
 extern char const ExeTypeStr[];